+++ superwall/src/main/java/com/superwall/sdk/store/products/GooglePlayProductsFetcher.kt
@@ -264,7 +264,6 @@ open class GooglePlayProductsFetcher(
         println("!!! onPurchasesUpdated $p0 $p1")
     }
 
-
     override suspend fun products(
         identifiers: Set<String>
     ): Set<StoreProduct> {