+++ CHANGELOG.md
@@ -2,13 +2,26 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall-me/Superwall-Android/releases) on GitHub.
 
+## 1.0.0-alpha.34
+
+### Breaking Changes
+
+- Changes `Superwall.instance.getUserAttributes()` to `Superwall.instance.userAttributes`.
+
+### Enhancements
+
+- SW-2665: Adds `Superwall.instance.userId` so you can access the current user's id.
+
+### Fixes
+
+- SW-2664: Fixes race condition between resetting and presenting paywalls.
+
 ## 1.0.0-alpha.33
 
 ### Fixes
 
 - Fixes issue where a user would be asked to enable notifications even if there weren't any attached 
-to the paywall. 
-
+to the paywall.
 
 ## 1.0.0-alpha.32